Changes the cli table format to be fancy
When stdout is not a tty, table style is not colorful

Work started to make it easier to deal with the wide tables.
go-pretty allows for fancier wrapping, truncating tables, hiding rows.

Paused work here, because this works, and we need to think about how to
dynamically downsize, as it is not immediately clear how to introspect
the table row length right before rendering.