Custom client types
As part of the ansible inventory 'escape hatch' from aerolab, we have been
manually overriding the 'ClientType' label to use for ansible roles.

This commit turns the 'ClientType' escape hatch into a first class citizen by
allowing 'none' type clusters to have a '--type-override' that labels the
cluster with a 'custom' role.

After deploying a cluster with the 'custom' role, the `aerolab inventory
ansible` command will add the instances in that cluster to that host group.

```json
{
  "_meta": {
    "hostvars": {
      "172.17.0.3": {
        "aerolab_cluster": "franz",
        "ansible_host": "172.17.0.3",
        "ansible_user": "root",
        "instance_id": "4eded405d881",
        "node_name": "testclient-1"
      }
    }
  },
  "kafka": {
    "hosts": [
      "172.17.0.3"
    ]
  }
}
```